Rename `/heccbot` to `/news`
